+++ SalesforceOrgManager/SalesforceOrgManager/View/BoxCreazioneTrigger.cs
@@ -17,6 +17,16 @@ namespace SalesforceOrgManager.View
         }
         private void btnCreateItem_Click(object sender, EventArgs e)
         {
+            if (String.IsNullOrWhiteSpace(txtItemName.Text))
+            {
+                MessageBox.Show("Name cannnot be empty", "Trigger creation", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+            if (cmbObjects.SelectedIndex == -1)
+            {
+                MessageBox.Show("Trigger must be related to a given object", "Trigger creation", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
             Program.createApexTriggerAsync(this);
         }
     }